reorg files and add formatter
