+++ package.json
@@ -3,11 +3,13 @@
   "version": "0.1.0",
   "description": "Example project for autonomys-agents",
   "type": "module",
-  "main": "dist/index.js",
+  "main": "dist/src/index.js",
   "scripts": {
     "build": "tsc",
-    "start": "npx tsx index.ts",
-    "dev": "tsx --no-cache --watch index.ts",
+    "start": "npx tsx src/index.ts",
+    "dev": "tsx --no-cache --watch src/index.ts",
+    "format": "prettier --write \"src/**/*.ts\"",
+    "format:check": "prettier --check \"src/**/*.ts\"",
     "create-character": "tsx scripts/create-character.ts",
     "generate-certs": "tsx scripts/generate-certs.ts"
   },
@@ -16,6 +18,7 @@
     "yargs": "^17.7.2"
   },
   "devDependencies": {
+    "prettier": "^3.5.3",
     "tsx": "^4.7.1",
     "typescript": "^5.3.3"
   },
