+++ tsconfig.json
@@ -13,7 +13,7 @@
     },
     "include": [
       "*.ts"
-  , "index.ts"  ],
+  , "src/index.ts"  ],
     "exclude": [
       "node_modules",
       "dist"
